fix(clients): reuse one HTTP connection per WikipediaClient instance, and align the attempt table's PK type with its join partner
LML#1192 review round 6, pass 3, A5: get_summary opened a fresh httpx.AsyncClient on every call, including retries within one call and repeat calls on the same WikipediaClient instance -- both real callers (the background warm, up to MAX_CANDIDATES_PER_WARM times per warm; the offline drain, once per artist across its entire session, on ONE WikipediaClient constructed for the whole run) hold a single instance and call get_summary repeatedly on it, so this was pure avoidable overhead: measured 3.45ms of event-loop-blocking CPU per fetch, ~30k avoidable TLS handshakes across a full drain. WikipediaClient now holds a lazily-built, per-instance httpx.AsyncClient behind wxyc_fastapi.http.async_singleton (the LML#241/#242 FD-leak-race fix), mirroring clients/streaming/base.py's identical adoption, released via a new aclose(). Verified live: patching httpx.AsyncClient's constructor and driving two get_summary calls on one instance now confirms exactly one construction, where it was two before the fix. Because both real callers already construct exactly one WikipediaClient for their whole scope (confirmed for the drain: constructed once in run_drain, passed as a parameter to every process_candidate call), this closes the gap with no call-site changes required anywhere.

A6: entity/artist_wikipedia_bio_attempt.py declared discogs_artist_id INTEGER, while the sibling table it's joined against (lml_cache.artist_wikipedia_bio, and the discogs-cache artist.id it both ultimately key off) declares BIGINT -- a mismatched join-key type is a latent index/plan hazard even where Postgres's implicit cast keeps today's queries correct. Aligned to BIGINT in both the runtime DDL and its .sql sidecar. No migration needed: this table has never held prod rows.

+++ entity/artist_wikipedia_bio_attempt.py
@@ -63,7 +63,7 @@ logger = logging.getLogger(__name__)
 
 _DDL_TABLE = """\
 CREATE TABLE IF NOT EXISTS lml_cache.artist_wikipedia_bio_attempt (
-    discogs_artist_id INTEGER PRIMARY KEY,
+    discogs_artist_id BIGINT PRIMARY KEY,
     attempted_at TIMESTAMPTZ NOT NULL DEFAULT now(),
     outcome TEXT NOT NULL
 )\
